% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hitters}
\alias{hitters}
\title{Baseball hitters career and 1986 season total statistics}
\format{
A \emph{263} by \emph{16} matrix. The variables are centered
to zero mean and standardized to unit variance.
\describe{
\item{TAB_86}{times at bat in 1986}
\item{HIT_86}{hits in 1986}
\item{HR_86}{home runs in 1986}
\item{RUN_86}{runs in 1986}
\item{RB_86}{runs batted-in in 1986}
\item{WAL_86}{walks in 1986}
\item{WAL}{walks during his career}
\item{PO_86}{put outs in 1986}
\item{ASS_86}{assists in 1986}
\item{ERR_86}{errors in 1986}
\item{YC}{years in the major leagues}
\item{TAB}{times at bat during his career}
\item{HIT}{hits during his career}
\item{HR}{home runs during his career}
\item{RUN}{runs during his career}
\item{RUNB}{runs batted-in during his career}
}
}
\source{
\url{http://lib.stat.cmu.edu/datasets/baseball.data}
}
\usage{
data(hitters)
}
\description{
16 statistics of 263 major league hitters some of the
overall career and others relative to the 1986 season. Available at StatLib.
The matrix has a block structure, defined by season offensive play, career
offensive play and season defensive play.
}
\keyword{datasets}
