% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hitters_labels}
\alias{hitters_labels}
\title{Baseball hitters statistics labels reference table}
\format{
A dataframe with columns
\describe{
\item{short.namethe}{labels ised in plots and tables}
\item{label}{the explanatory name}
\item{type}{whether offensive year 1986, defensive year 86 or offensive over the whole career}
}
}
\source{
\url{http://lib.stat.cmu.edu/datasets/baseball.data}
}
\usage{
data(hitters_labels)
}
\description{
This data frame provides descriptive labels for the variables in the
hitters datasets matching the short labels.
}
\keyword{datasets}
