% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holzinger}
\alias{holzinger}
\title{Holzinger student's ability data}
\format{
A \emph{145} by \emph{12} matrix. The variables are centered to zero mean.

\strong{spatial} (SPL)
\describe{
\item{visual}{Visual perception test, a nonlanguage multiple-choice test of spatial relations }
\item{cubes}{Cubes test, spatial relations}
\item{flags}{Lozenges test, a visual imagery test in two or three dimensions}
}
\strong{verbal} (VBL)
\describe{
\item{paragraph}{Paragraph comprehension test, comprehension as measured by completion and multiple-choice questions}
\item{sentence}{Sentence completion test, a multiple-choice test in which “correct” answers reflect good judgment on the part of the subject}
\item{wordm}{Word meaning test, a multiple-choice vocabulary test}
}
\strong{speed} (SPD)
\describe{
\item{addition}{Addition test, speed of adding pairs of one-digit numbers}
\item{counting}{Counting groups of dots test, 4–7 dots, arranged in random patterns to be counted by subject}
\item{straight}{Straight and curved capitals test, a series of capital letters to be distinguished between those composed of straight lines only and those containing curved lines}
}
\strong{mathematical} (MTH)
\describe{
\item{deduct}{Deduction test, logical deduction test using the symbols and the letters}
\item{numeric}{Numerical puzzles test, a numerical deduction test, the object being to supply four numbers which will produce four given answers employing the operations of addition, multiplication, or division}
\item{series}{Series completion test, from a series of five numbers, the subject is supposed to deduce the rule of procedure from one number to the next and thus supply the sixth number in the series}
}
}
\source{
The complete set of data is included, for example in the R package MBESS.
}
\usage{
data(holzinger)
}
\description{
Holzinger and Swineford (1939) data is widely cited. This dataset
includes only the 12 tests on the Grant-White School students used in Ferrara et al .
}
\references{
Holzinger, K. J. and Swineford, F. A. (1939). A study in factor analysis: The stability of a bi-factor solution. Supplementary Education Monographs, 48. University of Chicago.

Ferrara, C., Martella, F., and Vichi, M. (2019). Probabilistic disjoint principal component analysis. Multivariate Behavioral Research, 54(1):47–61.
}
\keyword{datasets}
