% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holzinger_labels}
\alias{holzinger_labels}
\title{Holzinger student's ability data labels reference table}
\format{
A dataframe with columns
\describe{
\item{Test}{Name of each test}
\item{ Ability}{Short label for the abilities}
\item{Long_Ability}{Long name of th abilities}
}
}
\description{
This data frame provides descriptive labels for the variables in the
Holzinger student's ability data matching the short labels used.
}
\keyword{datasets}
