% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsspca.R
\name{lsspca}
\alias{lsspca}
\title{Computes LS SPCA components using different variable selection algorithms}
\usage{
lsspca(
  X,
  alpha = 0.95,
  ncomps = 4,
  spcaMethod = "u",
  variableSelection = c("exhaustive", "seqrep", "backward", "forward", "lasso"),
  scalex = FALSE,
  maxcard = 0,
  force.in = NULL,
  force.out = NULL,
  selectfromthese = NULL,
  lsspca_forLasso = TRUE,
  lasso_penalty = 0.5
)
}
\arguments{
\item{X}{The data matrix.}

\item{alpha}{Real in [0,1]. percentage of variance of the PCs explained by the sparse component.}

\item{ncomps}{number of components to compute}

\item{spcaMethod}{character vector how LS SPCA components are computed:
"u" for uncorrelated, "c" for correlated and "p" for projection.
If only one value, the same method is used for all components.}

\item{variableSelection}{how the variables for each components are selected
'seqrep' stepwise, 'exhaustive' all subsets 'backward', 'forward', 'lasso'}

\item{scalex}{= FALSE, whether to scale the variables to unit variance. Variables are
scaled to zero mean (if needed) even if scaleX = FALSE}

\item{maxcard}{a vector or an integer. Missing values filled with last value.}

\item{force.in}{NULL or list of indeces that must be in component. not for lasso. [NULL]}

\item{force.out}{NULL or list of indeces cannot be in component. [NULL]}

\item{selectfromthese}{NULL or list of indeces from which model chosen. [NULL]}

\item{lsspca_forLasso}{use lsspca with indeces selected with lasso or just the lasso regression}

\item{lasso_penalty}{real between 0 and 1, , 0-> ridge regression, 1 -> lasso}
}
\value{
a list  with vif for each component
}
\description{
For each component, the variables are selected so as to explain
a percentage \emph{alpha} of the vexp by the corresponding principal component.
\emph{ind_blocks} is a list containing the indeces for each component,
}
\author{
Giovanni Merola
}
