% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsspca_blocked.R
\name{lsspca_blocked}
\alias{lsspca_blocked}
\title{For each component, the variables are selected so as to explain
a percentage \emph{alpha} of the vexp by the corresponding principal component.
\emph{blocks_list} is a list containing the indeies for each component,
Subsets can be overlapping and need not be exhaustive.}
\usage{
lsspca_blocked(X, alpha = 0.95, blocks_list = list(),
ncomps_per_block = 1, blocks_names = NA,
maxcard = 0, mincard = 1,
spcaMethod = c("u", "c", "p"), scalex = FALSE,
variableSelection = c("exhaustive", "seqrep", "backward", "forward", "lasso"),
lsspca_forLasso = FALSE, lasso_penalty = 0.5, rtn_all_spca = FALSE)
}
\arguments{
\item{X}{the data matrix.}

\item{alpha}{Real (or vector) in [0,1]. percentage of variance of the PCs explained by the sparse component.}

\item{blocks_list}{a list of indices or a vector or factor of codes for each block}

\item{ncomps_per_block}{number of components per block, integer or vector of. Default = 1.}

\item{blocks_names}{names of each block, if possible names are taken for the names of blocks_list}

\item{maxcard}{a vector or an integer of maximal cardinality for each block, if USPCA selected
maxcard[j] cannot be < j}

\item{mincard}{a vector or an integer}

\item{spcaMethod}{char how lsspca is computed "u" = USPCA (default), "c" = CSPCA, "p" = PSPCA}

\item{scalex}{Logical, if TRUE variables are scaled to unit variance.default  FALSE
Variables are automatically centered to zero if they aren't already.}

\item{variableSelection}{how the variables in each components are selected
"exhaustive" = all subsets, "seqrep" = stepwise, "backward", "forward", "lasso".
See documentation in packages leaps, and elasticnet for lasso.}

\item{lsspca_forLasso}{use lsspca with indeies selected with lasso, otherwise
just the lasso regression}

\item{lasso_penalty}{real between 0 and 1, , 0-> ridge regression, 1 -> lasso}

\item{rtn_all_spca}{logical, should all the block spca objects be returned, dafault FALSE.}
}
\value{
an a list
}
\description{
For each component, the variables are selected so as to explain
a percentage \emph{alpha} of the vexp by the corresponding principal component.
\emph{blocks_list} is a list containing the indeies for each component,
Subsets can be overlapping and need not be exhaustive.
}
\seealso{
\code{\link{lsspca}}
}
