% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makevexp.R
\name{makevexp}
\alias{makevexp}
\title{Utility for computing the variance explained by a set of components}
\usage{
makevexp(scores, X, unc = FALSE)
}
\arguments{
\item{scores, }{a matrix of components scores}

\item{X}{the data matrix.}

\item{unc}{logical, are the components orthogonal.}
}
\value{
a vector with the variance explained by each component
}
\description{
Utility for computing the variance explained by a set of components
}
\examples{
{
 if(FALSE){
   A = matrix(c(rep(1, 16), rep(c(-1, 0.5, 1, 0.5), 4)), 16, 2)
   sc = hitters \%*\% A
   makevexp(sc, hitters, FALSE)

 ## this is wrong
   makevexp(sc, hitters, TRUE)

 ## because the components are not orthogonal
   cor(sc)
 }
}
}
\author{
Giovanni Merola
}
