% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_spca}
\alias{print_spca}
\title{\strong{Prints the sparse loadings from an spca object}}
\usage{
print_spca(
  obj,
  contributions = TRUE,
  only.nonzero = TRUE,
  digits = 1,
  threshold = 0.001,
  prn = TRUE,
  rtn = FALSE
)
}
\arguments{
\item{obj}{An spca object.}

\item{contributions}{Logical: should the loadings be standardised to unit \eqn{L_1}
norm (and printed as percentage contributions)?}

\item{only.nonzero}{Logical: if = TRUE only the nonzero loadings are printed.
otherwise all loadings are printed.}

\item{digits}{Integer: number of decimal figures.}

\item{threshold}{Value below which loadings are considered zero and not
printed.}

\item{prn}{Logical: should the table be printed?}

\item{rtn}{Logical: should the formatted (text) table be returned?}
}
\value{
If rtn = TRUE, it returns a text table formatted as specified by the
arguments.
}
\description{
Prints sparse loadings omitting the zero ones and giving the cumulative
variance explained.
}
\seealso{
Examples in \link{lsspca}.
}
